Bug 3148 : Database revision to make zebraqueue.biblio_auth_number as big as auth_header.authid.
If you've never truncated your zebraqueue table, you might want to
update your index and truncate it before applying this patch and running updatedatabase.

+++ installer/data/mysql/kohastructure.sql
@@ -2115,7 +2115,7 @@ CREATE TABLE `z3950servers` (
 DROP TABLE IF EXISTS `zebraqueue`;
 CREATE TABLE `zebraqueue` (
   `id` int(11) NOT NULL auto_increment,
-  `biblio_auth_number` int(11) NOT NULL default '0',
+  `biblio_auth_number` bigint(20) unsigned NOT NULL default '0',
   `operation` char(20) NOT NULL default '',
   `server` char(20) NOT NULL default '',
   `done` int(11) NOT NULL default '0',
